refactor: move EnumExtensions to Extensions folder
