+++ src/Foundatio.RabbitMQ/Messaging/DelayedRetryType.cs
@@ -1,5 +1,3 @@
-using System;
-using System.Reflection;
 using System.Runtime.Serialization;
 
 namespace Foundatio.Messaging;
@@ -28,14 +26,4 @@ public enum DelayedRetryType
     /// </summary>
     [EnumMember(Value = "disabled")]
     Disabled
-}
-
-internal static class EnumExtensions
-{
-    public static string ToEnumString<T>(this T value) where T : struct, Enum
-    {
-        var member = typeof(T).GetField(value.ToString()!);
-        var attribute = member?.GetCustomAttribute<EnumMemberAttribute>();
-        return attribute?.Value ?? value.ToString()!;
-    }
 }
